+++ app/models/block_size_multiplier.rb
@@ -6,7 +6,7 @@ class BlockSizeMultiplier < ApplicationRecord
   # Concerns
   include Deletable
 
-  # Named Scopes
+  # Scopes
 
   # Model Validation
   validates :user_id, :project_id, :randomization_scheme_id, presence: true